feat: enhance user and chat schemas with OpenAPI examples
- Updated LoginResponse and UserResponse schemas to include OpenAPI examples for better documentation.
- Refactored S3StorageProvider to handle file uploads more robustly, including deletion of old files.
- Expanded Swagger documentation to include new chat and message endpoints, enhancing API usability.
- Implemented logging in DeleteChat case for better traceability.
- Modified ListChats to return ChatResponse instead of Chat, improving type safety.
- Updated ChatController to handle participant parsing more effectively during chat creation and updates.
- Added new schemas for message responses and chat parameters, ensuring consistent validation and documentation.
- Refined message handling in CreateMessage and UpdateMessage cases to return structured responses.
- Introduced new DTOs for chat and message operations, improving code organization and clarity.

+++ src/infra/providers/storage/S3StorageProvider.ts
@@ -42,19 +42,34 @@ export class S3StorageProvider implements IStorageProvider {
     mimeType: string;
     oldFileUrl: string;
   }): Promise<string> {
-    const bucketUrlPrefix = `https://${process.env.S3_BUCKET_NAME}.s3.amazonaws.com/`;
-    const key = oldFileUrl.replace(bucketUrlPrefix, '');
+
     const extension = path.extname(fileName);
-    const newKey = `${path.dirname(key)}/${uuidv4()}${extension}`;
+    let newKey: string;
 
-    await this.client.send(new PutObjectCommand({
-      Bucket: process.env.S3_BUCKET_NAME!,
-      Key: newKey,
-      Body: fileBuffer,
-      ContentType: mimeType
-    }));
+    if (oldFileUrl && oldFileUrl.trim() !== '') {
+      const url = new URL(oldFileUrl);
+      const key = url.pathname.startsWith('/') ? url.pathname.slice(1) : url.pathname;
+      newKey = `${path.dirname(key)}/${uuidv4()}${extension}`;
 
-    await this.deleteFile(oldFileUrl);
+      await this.client.send(new PutObjectCommand({
+        Bucket: process.env.S3_BUCKET_NAME!,
+        Key: newKey,
+        Body: fileBuffer,
+        ContentType: mimeType
+      }));
+      await this.deleteFile(oldFileUrl);
+
+    } else {
+      newKey = `${uuidv4()}${extension}`;
+
+
+      await this.client.send(new PutObjectCommand({
+        Bucket: process.env.S3_BUCKET_NAME!,
+        Key: newKey,
+        Body: fileBuffer,
+        ContentType: mimeType
+      }));
+    }
 
     return `https://${process.env.S3_BUCKET_NAME}.s3.amazonaws.com/${newKey}`;
   }
